+++ docs/news.rst
@@ -8,6 +8,7 @@ Changes
 * Added support for json and csv output.
 * Improved command line support for API calls.
 * Command line parameters can now override config file.
+* Added support for initiating callback.
 
 0.1
 ---